CoreProtect and other Quality of Life Improvements
Added CoreProtect to break tasks.
CoreProtect is a soft-depency.
Changed PlaceholderAPI as a soft-dependency instead of required.
Removed Vault as a soft-dependency. Vault is not used in this plugin.
Added more smartitude for checking task IDs when used in lore sections. No more NPE vomits.
Added a togo variable for quest lore.